Last Updates

+++ src/pages/Reports.js
@@ -1,14 +1,20 @@
 import React, { useEffect, useState } from 'react';
 import * as S from './styledPages';
+import { useHistory } from 'react-router-dom';
 
 export default function Reports() {
+  const history = useHistory();
   const [ repositories, setRepositories ] = useState([]);
   
   useEffect(() => {
     let repositoriesName = localStorage.getItem('repositoriesName');
+    if(repositoriesName !== null) {
     repositoriesName = JSON.parse(repositoriesName);
     setRepositories(repositoriesName);
     localStorage.clear();
+    } else {
+      history.push(`/`);
+    }
   }, []);
 
   return (
